+++ src/routes/(app)/room/[room]/authors.ts
@@ -2,11 +2,9 @@ import type { Profile } from '$lib/db/users';
 import { getUserProfileById } from '$lib/db/users';
 import type { Supabase } from '$lib/supabase';
 import { getSession } from '$lib/supabase';
-import { decryptRoomSessionKey, InboundSession, OutboundSession } from 'moe';
+import { InboundSession, OutboundSession } from 'moe';
 import type { AuthoredMessage } from '$lib/db/messages';
-import { olmAccount, raise } from '$lib/utils';
-import { get as getStore } from 'svelte/store';
-import { buildOutboundSession } from '$lib/e2ee';
+import { buildOutboundSession, getDecryptedSessionKey } from '$lib/e2ee';
 
 const authors: { [key: string]: Profile } = {};
 
@@ -24,65 +22,13 @@ export async function get(supabase: Supabase, id: string): Promise<Profile> {
 }
 
 async function getInboundSessionViaOlmSessionKey(supabase: Supabase, selfId: string, authorId: string, roomId: string) {
-    const { data: sessKey, error: sessKeyErr } = await supabase
-        .from('room_session_keys')
-        .select('*')
-        .eq('room_id', roomId)
-        .eq('key_of', authorId)
-        .eq('key_for', selfId)
-        .maybeSingle();
-    if (sessKeyErr) {
-        throw sessKeyErr;
+    const sessionKey = await getDecryptedSessionKey(supabase, selfId, authorId, roomId);
+    if (sessionKey === null) {
+        return null;
     }
-    if (sessKey === null) {
-        // TODO: this is the case where they haven't given us a key
-        // they should give us a key when inviting(/adding?) us to a room
-        return null
-    }
-
-    const { data, error: idKeyErr } = await supabase
-        .from('user_identity_keys')
-        .select('curve25519')
-        .eq('id', authorId)
-        .single();
-
-    if (idKeyErr) {
-        throw idKeyErr;
-    }
-
-    const ciphertext = sessKey.key!;
-    const encryptorIdentityKey = data.curve25519;
-    const userAccount = getStore(olmAccount) ?? raise('olm account not initialized');
-    const sessionKey = decryptRoomSessionKey(userAccount, encryptorIdentityKey, ciphertext);
     localStorage.setItem(`${roomId}:${authorId}:sessionKey`, sessionKey);
     // sessionKeys[`${roomId}:${authorId}`] = inbound;
     return new InboundSession(sessionKey);
-}
-
-export async function getInboundSession(supabase: Supabase, authorId: string, roomId: string) {
-    const cached = sessionKeys[`${roomId}:${authorId}`];
-    if (cached) return cached;
-
-    console.info('cache miss for inbound session; fetching from supabase', authorId, roomId);
-    const key = await supabase
-        .from('room_members')
-        .select('*')
-        .eq('member_id', authorId)
-        .eq('room_id', roomId)
-        .limit(1)
-        .single();
-
-    if (key.error) {
-        throw key.error;
-    }
-
-    const sessionKey = key.data.session_key;
-    if (sessionKey === null) {
-        throw Error('attempted to get inbound session for a room without end-to-end encryption');
-    }
-
-    sessionKeys[`${roomId}:${authorId}`] = new InboundSession(sessionKey);
-    return sessionKeys[`${roomId}:${authorId}`];
 }
 
 export async function decryptMessage(supabase: Supabase, message: AuthoredMessage) {
